requirements.txt
fastmcp>=2.3.0
google-genai>=1.5.0
uvicorn>=0.32.0
pillow>=11.0.0
